Move tests into specific unit + integration dirs
