+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "npm run lint && npm run test:integration && npm run coverage",
-    "test:unit": "tape test/unit-tests.js | tap-spec",
-    "test:integration": "tape test/integration-tests.js | tap-spec",
+    "test:unit": "tape test/unit/*-test.js | tap-spec",
+    "test:integration": "tape test/integration/integration-tests.js | tap-spec",
     "coverage": "nyc --reporter=lcov --reporter=text-summary npm run test:unit",
     "lint": "eslint . --fix --ignore-pattern node_modules"
   },
