+++ test/integration/mock/_optional/shared/.arc-config
