+++ test/integration/mock/_optional/shared/package.json
