+++ test/integration/mock/_optional/views/package.json
