+++ test/integration/mock/public/static.json
