+++ test/integration/mock/src/events/just-being-in-nature/package.json
