+++ test/integration/mock/src/http/delete-badness_in_life/.arc-config
