+++ test/integration/mock/src/http/get-index/package.json
