+++ test/integration/mock/src/http/get-memories/.arc-config
