+++ test/integration/mock/src/http/get-memories/requirements.txt
