+++ test/integration/mock/src/http/get-memories/setup.cfg
