+++ test/integration/mock/src/http/put-on_your_boots/package.json
