+++ test/integration/mock/src/scheduled/hikes-with-friends/index.js
