+++ test/integration/mock/src/scheduled/hikes-with-friends/package.json
