+++ test/integration/mock/src/tables/trails-insert/index.js
