+++ test/integration/mock/src/tables/trails-insert/package.json
